+++ src/mpt/db.go
@@ -1,4 +1,4 @@
-package main
+package mpt
 
 import (
 	"github.com/syndtr/goleveldb/leveldb"
