+++ src/mpt/empty_test.go
@@ -1,4 +1,4 @@
-package main
+package mpt
 
 import (
 	"testing"
