+++ src/mpt/trie_test.go
@@ -1,4 +1,4 @@
-package main
+package mpt
 
 import (
 	"fmt"
